# openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Expenses API
  description: |
    REST API for managing restaurant expenses.

    All `/expenses` endpoints require a Bearer token in the `Authorization` header.
    The token must be a valid UUID.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

security:
  - bearerAuth: []

tags:
  - name: expenses
    description: Expense management

paths:
  /expenses/:
    get:
      tags: [expenses]
      summary: List expenses
      description: Returns a paginated list of expenses with optional filters.
      operationId: listExpenses
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - name: text
          in: query
          description: Free-text search filter.
          schema:
            type: string
        - name: supplier
          in: query
          description: Filter by supplier name.
          schema:
            type: string
        - name: from
          in: query
          description: Include expenses issued at or after this timestamp (RFC3339).
          schema:
            type: string
            format: date-time
            example: '2026-01-15T10:30:00Z'
        - name: to
          in: query
          description: Include expenses issued at or before this timestamp (RFC3339).
          schema:
            type: string
            format: date-time
            example: '2026-01-31T23:59:59Z'
      responses:
        '200':
          description: Paginated list of expenses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExpensesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags: [expenses]
      summary: Create expense
      description: Creates a new expense record.
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
      responses:
        '201':
          description: Expense created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /expenses/{id}/:
    get:
      tags: [expenses]
      summary: Get expense
      description: Returns a single expense by ID.
      operationId: getExpense
      parameters:
        - $ref: '#/components/parameters/expenseId'
      responses:
        '200':
          description: Expense found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

    put:
      tags: [expenses]
      summary: Update expense
      description: Updates an existing expense by ID.
      operationId: updateExpense
      parameters:
        - $ref: '#/components/parameters/expenseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
      responses:
        '200':
          description: Expense updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

    delete:
      tags: [expenses]
      summary: Delete expense
      description: Deletes an expense by ID.
      operationId: deleteExpense
      parameters:
        - $ref: '#/components/parameters/expenseId'
      responses:
        '204':
          description: Expense deleted.
        '400':
          $ref: '#/components/responses/BadRequest'

  /expenses/import/:
    post:
      tags: [expenses]
      summary: Import expenses
      description: |
        Accepts a batch of expenses for asynchronous import.
        Returns immediately with `202 Accepted`; processing continues in the background.
      operationId: importExpenses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/ExpenseRequest'
      responses:
        '202':
          description: Import accepted for processing.
        '400':
          $ref: '#/components/responses/BadRequest'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: Bearer token containing a valid UUID.

  parameters:
    expenseId:
      name: id
      in: path
      required: true
      description: Expense identifier (UUID).
      schema:
        type: string
        format: uuid

    page:
      name: page
      in: query
      description: Page number (1-based). Defaults to 1.
      schema:
        type: integer
        minimum: 1
        default: 1

    pageSize:
      name: page_size
      in: query
      description: Number of items per page. Defaults to 25, maximum 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25

  schemas:
    ExpenseRequest:
      type: object
      required:
        - supplier
        - issuedAt
        - qty
        - costPerUnit
        - category
      properties:
        supplier:
          type: string
          minLength: 1
        issuedAt:
          type: string
          format: date-time
          description: RFC3339 timestamp.
          example: '2026-01-15T10:30:00Z'
        description:
          type: string
        qty:
          type: integer
          minimum: 1
        costPerUnit:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          description: Unit cost with at most two decimal places.
          example: 12.50
        category:
          type: string
          minLength: 1

    ExpenseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        supplier:
          type: string
        issuedAt:
          type: string
          format: date-time
        description:
          type: string
        qty:
          type: integer
        costPerUnit:
          type: number
          format: double
        category:
          type: string

    PaginatedExpensesResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseResponse'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: validation error

  responses:
    BadRequest:
      description: Invalid request, validation failure, or authorization error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'